requirements.txt
httpx>=0.24.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
bilibili-api-python>=17.3.0
